+++ sync-for-payroll/src/codatsyncpayroll/models/operations/list_journal_entries.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import errormessage as shared_errormessage
-from ..shared import journalentries as shared_journalentries
+from ...models.shared import journalentries as shared_journalentries
 from typing import Optional
 
 
@@ -28,13 +27,11 @@ class ListJournalEntriesRequest:
 class ListJournalEntriesResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your `query` parameter was not correctly formed"""
     journal_entries: Optional[shared_journalentries.JournalEntries] = dataclasses.field(default=None)
     r"""Success"""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""