+++ sync-for-payroll/src/codatsyncpayroll/models/shared/pushvalidationinfo.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 import dataclasses
-from ..shared import pushfieldvalidation as shared_pushfieldvalidation
+from .pushfieldvalidation import PushFieldValidation
 from codatsyncpayroll import utils
 from dataclasses_json import Undefined, dataclass_json
 from typing import List, Optional
@@ -11,7 +11,7 @@ from typing import List, Optional
 @dataclass_json(undefined=Undefined.EXCLUDE)
 @dataclasses.dataclass
 class PushValidationInfo:
-    information: Optional[List[shared_pushfieldvalidation.PushFieldValidation]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('information') }})
-    warnings: Optional[List[shared_pushfieldvalidation.PushFieldValidation]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('warnings') }})
+    information: Optional[List[PushFieldValidation]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('information') }})
+    warnings: Optional[List[PushFieldValidation]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('warnings') }})